+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/festivaldetail/FestivalDetailViewModel.kt
@@ -5,17 +5,20 @@ import androidx.lifecycle.viewModelScope
 import com.festago.festago.common.analytics.AnalyticsHelper
 import com.festago.festago.common.analytics.logNetworkFailure
 import com.festago.festago.domain.model.artist.Artist
+import com.festago.festago.domain.model.bookmark.BookmarkType
 import com.festago.festago.domain.model.festival.FestivalDetail
 import com.festago.festago.domain.model.school.School
 import com.festago.festago.domain.model.stage.Stage
 import com.festago.festago.domain.repository.BookmarkRepository
 import com.festago.festago.domain.repository.FestivalRepository
+import com.festago.festago.domain.repository.UserRepository
 import com.festago.festago.presentation.ui.festivaldetail.FestivalDetailEvent.FailedToFetchBookmarkList
 import com.festago.festago.presentation.ui.festivaldetail.uiState.ArtistItemUiState
 import com.festago.festago.presentation.ui.festivaldetail.uiState.FestivalDetailUiState
 import com.festago.festago.presentation.ui.festivaldetail.uiState.FestivalUiState
 import com.festago.festago.presentation.ui.festivaldetail.uiState.StageItemUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharedFlow
@@ -29,6 +32,7 @@ import javax.inject.Inject
 class FestivalDetailViewModel @Inject constructor(
     private val festivalRepository: FestivalRepository,
     private val bookmarkRepository: BookmarkRepository,
+    private val userRepository: UserRepository,
     private val analyticsHelper: AnalyticsHelper,
 ) : ViewModel() {
 
@@ -43,15 +47,18 @@ class FestivalDetailViewModel @Inject constructor(
 
         viewModelScope.launch {
             val deferredFestivalDetail =
-                festivalRepository.loadFestivalDetail(festivalId, delayTimeMillis)
-            val deferredBookmarks = bookmarkRepository.getFestivalBookmarkIds()
+                async { festivalRepository.loadFestivalDetail(festivalId, delayTimeMillis) }
 
             runCatching {
-                val festivalDetail = deferredFestivalDetail.getOrThrow()
-                val festivalBookmarkIds = deferredBookmarks.getOrThrow()
+                val festivalDetail = deferredFestivalDetail.await().getOrThrow()
 
-                _uiState.value = festivalDetail.toSuccessUiState()
-                    .copy(bookmarked = festivalBookmarkIds.contains(festivalId))
+                val isBookmarked = if (userRepository.isSigned()) {
+                    bookmarkRepository.isBookmarked(festivalId, BookmarkType.FESTIVAL)
+                } else {
+                    false
+                }
+
+                _uiState.value = festivalDetail.toSuccessUiState(isBookmarked)
             }.onFailure {
                 _uiState.value = FestivalDetailUiState.Error { festivalId ->
                     _uiState.value = FestivalDetailUiState.Loading
@@ -65,25 +72,31 @@ class FestivalDetailViewModel @Inject constructor(
         }
     }
 
-    private fun FestivalDetail.toSuccessUiState() = FestivalDetailUiState.Success(
-        festival = FestivalUiState(
-            id = id,
-            name = name,
-            startDate = startDate,
-            endDate = endDate,
-            posterImageUrl = posterImageUrl,
-            school = school,
-            onSchoolClick = ::showSchoolDetail,
-            socialMedias = socialMedias,
-        ),
-        bookmarked = false,
-        stages = stages.map { it.toUiState() },
-        onBookmarkClick = { festivalId -> toggleFestivalBookmark(festivalId) },
-    )
+    private fun FestivalDetail.toSuccessUiState(isBookmarked: Boolean) =
+        FestivalDetailUiState.Success(
+            festival = FestivalUiState(
+                id = id,
+                name = name,
+                startDate = startDate,
+                endDate = endDate,
+                posterImageUrl = posterImageUrl,
+                school = school,
+                onSchoolClick = ::showSchoolDetail,
+                socialMedias = socialMedias,
+            ),
+            bookmarked = isBookmarked,
+            stages = stages.map { it.toUiState() },
+            onBookmarkClick = { festivalId -> toggleFestivalBookmark(festivalId) },
+        )
 
     private fun toggleFestivalBookmark(festivalId: Long) {
         viewModelScope.launch {
             val uiState = _uiState.value as? FestivalDetailUiState.Success ?: return@launch
+
+            if (!userRepository.isSigned()) {
+                _event.emit(FailedToFetchBookmarkList("로그인이 필요합니다"))
+                return@launch
+            }
 
             if (uiState.bookmarked) {
                 bookmarkRepository.deleteFestivalBookmark(festivalId)
